Module 1: Package structure
Signed off by: Mauro Sarti mauro.sarti@globant.com
