+++ lib/ui/screen/home/HomeScreen.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'HomeState.dart';
+
+class HomeScreen extends StatefulWidget {
+  const HomeScreen({super.key, required this.title});
+
+  final String title;
+
+  @override
+  State<HomeScreen> createState() => HomeState();
+}
